mini_api/requirements.txt
Django==5.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
